Import Models w/ Linked Duplicates
- Louron Core -

Updated the ModelImporter to check for linked duplicates by creating an unordered map of loaded nodes, which is keyed by the aiNode meshes that are included, and the value is a pair of data for the MeshFilterComponent, and the MeshRendererComponent.

Created GetTransform for Entity which basically just gets the transform without having to do the whole GetComponent<TransformComponent>() call.

Updated entry point to be more clearer with the macros for enabling and disabling console.

- Louron Editor -

Created test scene which demonstrates how a Model with linked duplicates can be automatically instanced - check the Rendering Stats to see how it is automatically instancing the geometry that are linked duplicates.

+++ Louron Core/src/Scene/Components/Mesh.cpp
@@ -306,7 +306,7 @@ namespace Louron {
 		}
 
 		// Apply new normals
-		SetNormals(new_normals, vertice_count);
+		SetNormals(new_normals, static_cast<GLuint>(vertice_count));
 
 		delete[] new_normals;
 	}
